back button moved to toolbar toolbar dropped from login activity (no functionality)
