+++ app/src/main/res/layout/login.xml
@@ -7,15 +7,6 @@
     android:layout_height="match_parent"
     tools:context="com.greenfox.fuchsit.zerdareader.activity.HelloWorldActivity">
 
-    <android.support.v7.widget.Toolbar
-        android:id="@+id/my_toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="?attr/actionBarSize"
-        android:background="?attr/colorPrimary"
-        android:elevation="4dp"
-        android:theme="@style/ThemeOverlay.AppCompat.ActionBar"
-        app:popupTheme="@style/ThemeOverlay.AppCompat.Light"/>
-
     <RelativeLayout
         android:orientation="vertical"
         android:layout_width="match_parent"
