+++ app/src/main/res/menu/my_toolbar_menu.xml
@@ -13,7 +13,14 @@
         android:id="@+id/favorite"
         android:icon="@drawable/ic_star"
         android:title="@string/favorite"
-        app:showAsAction="ifRoom"/>
+        app:showAsAction="always"/>
+
+    <item
+        android:id="@+id/back"
+        android:icon="@drawable/ic_back_button"
+        android:title="@string/back"
+        app:showAsAction="always"/>
+
 
     <item
         android:id="@+id/manage_subscriptions"
