+++ app/src/main/res/values/strings.xml
@@ -16,6 +16,7 @@
     </string>
     <string name="favorite">Favorite</string>
     <string name="refresh">Refresh</string>
+    <string name="back">Back</string>
     <string name="Settings">settings</string>
     <string name="manage_subscriptions">Manage subscriptions</string>
     <string name="settings">Settings</string>
